feat: 6주차 미션 완료

+++ 라미/server/src/main/java/umc/server/domain/review/entity/Review.java
@@ -25,7 +25,7 @@ public class Review extends BaseEntity {
     private String content;
 
     @Column(name = "rating", nullable = false)
-    private int rating;
+    private Float rating;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "member_id")
